added Landing mobile section and updated components used by it
